requirements.txt
Django==4.2.1
python-telegram-bot
mysqlclient


